requirements.txt
django_extensions
pillow
pip install stripe

pip install coverage
coverage run --source='.' manage.py test
coverage report

pip install pytest
pytest viewer/test_123.py

pip install selenium